+++ ext/FluxnetSimulationsExt.jl
@@ -1,6 +1,16 @@
 module FluxnetSimulationsExt
+import ClimaUtilities.TimeVaryingInputs:
+    TimeVaryingInput, LinearInterpolation, PeriodicCalendar
+import ClimaUtilities.TimeManager: ITime, date
+using Thermodynamics
+using Dates
+using DelimitedFiles
+using DocStringExtensions
+using Insolation
+import ClimaLand.Parameters as LP
 using ClimaLand
-import DelimitedFiles
+export prescribed_forcing_fluxnet,
+    set_fluxnet_ic!, get_comparison_data, get_data_dates, get_data_dt
 
 # Include site-specific configurations, as well as the default generic site.
 include("fluxnet_simulations/generic_site.jl")
@@ -10,4 +20,7 @@ include("fluxnet_simulations/US-NR1.jl")
 include("fluxnet_simulations/US-Var.jl")
 # export get_parameters, get_domain_info - TODO uncomment once we implement these
 
+include("fluxnet_simulations/data_processing.jl")
+include("fluxnet_simulations/forcing.jl")
+include("fluxnet_simulations/initial_conditions.jl")
 end # module FluxnetSimulationsExt